+++ vercel.json
@@ -14,5 +14,6 @@
     }
   ],
   "headers": [],
-  "buildCommand": "npm run build"
+  "buildCommand": "npm install --legacy-peer-deps && npm run build",
+  "installCommand": "npm install --legacy-peer-deps"
 }
